Moved Pimoroni sensors into own folder
