+++ sensor_modules/pimoroni/pimoroni_mcp9600.py
